Adding TypeRenkoSymetric
TypeRenkoSymetric: Custom renko version. It is quite similar to
traditional renko, but it changes the threshold when a renko
trend change direction. For example, if the brick size is 10 and the first
price is 95, the initial renko thresholds are 85 and 105. If a new price reached 105,
there is a new brick, at 95, and on traditional renko the new brick
thresholds are 85 and 115. It means, there are 10 points to increase to a
new brick, but 20 to deacrease. On TypeRenkoSymetric, it is kept 10
points on both size. It means, with a new price of 105, there is a new
brick at 105, but the new thresholds are 95 and 115 (10 points on both
sides).